requirement.txt
# - python=3.10.0 - pytorch-cuda=11.7
numpy==1.26.4
torch==2.4.0
torchvision==0.19.0
xformers==0.0.27.post2
torchaudio==2.4.0
setuptools
cmake
lit
accelerate==1.2.1
diffusers==0.32.2
transformers==4.47.1
einops
huggingface-hub==0.30.2
opencv-python
timm
safetensors
omegaconf
matplotlib
pynvml
scikit-image
Pillow
deepspeed==0.15.4